refactor(direct-control): extract notification verification
Move wrapper-level notification dismissal polling and identity verification into packages/civ7-direct-control/src/play/notifications/verification.ts while keeping the embedded App UI dismissal source owner and the public read/send wrappers in the facade.

Record the completed notification verification atom in the support modularization OpenSpec tasks and workstream records. This is local package and source relocation proof only, not runtime/live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- notification-dismissal.test.ts; bun run --cwd packages/cli test -- game/play/notification/dismiss.test.ts game/play/notification/dismiss-queue.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; bun run check:cli; bun run test:cli:play; git diff --check.
